Moved the parser to the root of the realm-browser
